chore: refactor and stop sorting keys
